pyproject.toml
[project]
name = "live-stt"
version = "0.1.0"
description = "Live Japanese speech-to-text with English translation"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24",
    "python-dotenv>=1.0",
]

[project.scripts]
live-stt = "live_stt:main"
